+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 9.3.0
+
+* Add `SETNX`, `ECHO`, `TYPE`, `EXPIRETIME`, and `PEXPIRETIME` commands
+* Add hashmap expiration commands (`HTTL`, `HEXPIRE`, etc)
+* Change `active_connections` to preempt reconnections
+
 ## 9.2.1
 
 * Fix docs.rs documentation features